Add automation for the chess puzzle
It does a depth-first search through the tree of possible board states.
As soon as it finds a set of moves that results in a queen cookie, it
makes them.

It assumes that all puzzles are solvable, which appears to be true both
from my experience the spading I've read.